Migrate network specific stuff to Network
Since it is not necessary for the DryRun, it must be specific to the Network so
I will move it back into that class.
